+++ client/src/styles/theme.ts
@@ -144,6 +144,8 @@ const color = {
     created: '#2E7D32',
     about: '#000000',
     setAbout: '#2D60B4',
+    setCreated: '#700370',
+    setPublished: '#6e6e6e',
   },
 
   advancedSearch: {