+++ client/CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file.
 
+## v1.63.0
+
+### Added
+
+- Added Objects and Works overlay button to objects/works included section ([#559](https://github.com/project-lux/lux-frontend/issues/559)).
+- Added support for Collections Created, Published, or Influenced By on P&G pages ([#837](https://github.com/project-lux/lux-frontend/issues/837)).
+
+### Changed
+
+- Changed the selected filters labeling ([#608](https://github.com/project-lux/lux-frontend/issues/608)).
+
+### Fixed
+
+- Fixed text alignment in results page tab buttons ([#520](https://github.com/project-lux/lux-frontend/issues/520)).
+- Fixed spacing of bar chart slider labels ([#674](https://github.com/project-lux/lux-frontend/issues/674)).
+- Fixed date facets labeling after user submission ([#509](https://github.com/project-lux/lux-frontend/issues/509)).
+
+### Removed
+
+- Removed Set option from Work Class in advanced search ([#735](https://github.com/project-lux/lux-frontend/issues/735)).
+
 ## v1.62.0
 
 ### Added